MLLM backend: refactor by moving MLLMModel to annif.lexical.mllm module
